+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Core: Fixed casing of Move.PassString
 * Viewer: allow parsing game strings from Engine output even if Mzinga thinks the moves are invalid
+* Perft: Allow passing in starting GameString 
 
 ## v0.10.7 ##